Initial commit
ICON Size 	-- Devices
60×60		-- iPhone & iPod Touch (Non-Retina)
76×76		-- iPad 2 & iPad mini (Non-Retina)
120×120		-- iPhone & iPod Touch (Retina)
152×152		-- iPad & iPad mini (Retina)
Note: Safari on iOS 7 doesn’t add any effects to these icons, however older versions of Safari will.
We don’t want older versions of Safari to apply effects to our icons, so making sure that the filenames end with "-precomposed.png".
Note: The icon that is the most appropriate size for the device is used. If no sizes attribute is set, the element’s size defaults to 60 x 60.