Initial commit
Specifying a Startup Image.
Note: On iOS, similar to native applications, we specified a startup image that is displayed while our web application launches. This is especially useful when our web application is offline.
By default, a screenshot of the web application the last time it was launched is used.